requirements.txt
# ───────────────────────────────────────────────
# Fraud Detection System — Python Dependencies
# ───────────────────────────────────────────────

# Core data science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Gradient Boosting (optional — can substitute with sklearn GBM)
# xgboost>=2.0.0        # uncomment for XGBoost support
# lightgbm>=4.0.0       # uncomment for LightGBM support

# Model serialisation
joblib>=1.3.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
tqdm>=4.65.0

# API & Web Server
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Testing & Rate Limiting
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
slowapi>=0.1.8
